Pull 0.57.7

+++ React/Modules/RCTLayoutAnimation.m
@@ -1,10 +1,8 @@
 /**
- * Copyright (c) 2015-present, Facebook, Inc.
- * All rights reserved.
+ * Copyright (c) Facebook, Inc. and its affiliates.
  *
- * This source code is licensed under the BSD-style license found in the
- * LICENSE file in the root directory of this source tree. An additional grant
- * of patent rights can be found in the PATENTS file in the same directory.
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
  */
 
 #import "RCTLayoutAnimation.h"
@@ -30,7 +28,7 @@ static UIViewAnimationOptions UIViewAnimationOptionsFromRCTAnimationType(RCTAnim
       // http://stackoverflow.com/questions/18870447/how-to-use-the-default-ios7-uianimation-curve
       return (UIViewAnimationOptions)(_currentKeyboardAnimationCurve << 16);
     default:
-      RCTLogError(@"Unsupported animation type %zd", type);
+      RCTLogError(@"Unsupported animation type %lld", (long long)type);
       return UIViewAnimationOptionCurveEaseInOut;
   }
 }
